chore(lint): azul-css clippy clean + core/layout autofix sweep
azul-css had 4 extreme-lint errors that ABORTED the CI Clippy job,
masking the rest: two doc-backtick items, one over-long first doc
paragraph, one explicit `.iter()` loop. Fixed by hand; `cargo clippy
-p azul-css --all-targets -- -D warnings` is clean.

With css unblocked, core+layout surfaced ~146 extreme-lint errors.
This commit carries clippy's OWN machine-applied portion (--fix:
~130 mechanical rewrites across 22 files — inline format args,
unneeded returns, char-slice `split`, fn-pointer hash without the
usize cast). The remaining ~110 (struct-name repetition, unnecessary
qualification, doc backticks/paragraphs, must_use, # Errors sections)
are hand-work tracked as the next lint chunk — Clippy CI stays red
with a reduced count until then.

+++ core/src/diff.rs
@@ -4218,7 +4218,7 @@ pub struct PreCascadeTransfers {
                 match img.get_data() {
                     DecodedImage::Callback(cb) => {
                         0xB0DE_CA11u32.hash(&mut h);
-                        (cb.callback.cb as usize).hash(&mut h);
+                        cb.callback.cb.hash(&mut h);
                         cb.refany.get_type_id().hash(&mut h);
                     }
                     _ => {
